+++ app/src/main/java/com/dandrzas/inertialsensors/data/SystemAlgorithm.java
@@ -8,12 +8,12 @@ public class SystemAlgorithm extends Observable implements IFOrientationAlgorith
     private float[] rollPitchYaw = new float[3];
     private float[] rotationMatrixNWURPY = new float[9];
     private float[] rotationMatrixRPYNWU = new float[9];
-    private float timestamp;
+    private long timestamp;
 
     public SystemAlgorithm() {
     }
 
-    public void calcOrientation(float[] systemRotationSensorVal, float timestamp) {
+    public void calcOrientation(float[] systemRotationSensorVal, long timestamp) {
         this.timestamp = timestamp;
         SensorManager.getRotationMatrixFromVector(rotationMatrixNWURPY, systemRotationSensorVal);
         float[] systemOrientation = new float[4];
@@ -64,7 +64,7 @@ public class SystemAlgorithm extends Observable implements IFOrientationAlgorith
     }
 
     @Override
-    public float getSampleTime() {
+    public long getSampleTime() {
         return timestamp;
     }
 }
